refactor(README) : model 패키지로 class 이동.
