+++ src/main/java/bridge/model/BridgeMaker.java
@@ -1,13 +1,11 @@
-package bridge;
+package bridge.model;
 
+import bridge.BridgeNumberGenerator;
 import bridge.constant.BridgeConstant;
 
 import java.util.LinkedList;
 import java.util.List;
 
-/**
- * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
- */
 public class BridgeMaker {
 
     private final BridgeNumberGenerator bridgeNumberGenerator;
